add missing pages: oauth2 / platform
